+++ src/main/java/com/spring/blackcat/common/code/FontWeightType.java
@@ -1,4 +1,4 @@
-package com.spring.blackcat.code;
+package com.spring.blackcat.common.code;
 
 // TODO: 2022-10-10 Enum으로 관리되는 것들이 DB에 저장될 때 대문자말고 camelCase로 저장되도록
 public enum FontWeightType {
