+++ src/main/java/com/spring/blackcat/common/code/Role.java
@@ -1,4 +1,4 @@
-package com.spring.blackcat.code;
+package com.spring.blackcat.common.code;
 
 public enum Role {
     ADMIN(Values.ADMIN), TATTOOIST(Values.TATTOOIST), BASIC(Values.BASIC);
