+++ src/main/java/com/spring/blackcat/likes/LikesRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.spring.blackcat.likes;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.spring.blackcat.code.PostType;
+import com.spring.blackcat.common.code.PostType;
 import com.spring.blackcat.likes.dto.LikesPostResDto;
 import com.spring.blackcat.likes.dto.LikesUserResDto;
 import org.springframework.data.domain.Page;
