+++ src/main/java/com/spring/blackcat/likes/dto/LikesPostResDto.java
@@ -1,6 +1,6 @@
 package com.spring.blackcat.likes.dto;
 
-import com.spring.blackcat.code.PostType;
+import com.spring.blackcat.common.code.PostType;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.NoArgsConstructor;
