+++ src/main/java/com/spring/blackcat/magazine/dto/MagazineTitleReqDto.java
