+++ src/test/java/com/spring/blackcat/likes/LikesRepositoryTest.java
@@ -4,9 +4,9 @@ import com.spring.blackcat.address.Address;
 import com.spring.blackcat.address.AddressRepository;
 import com.spring.blackcat.category.Category;
 import com.spring.blackcat.category.CategoryRepository;
-import com.spring.blackcat.code.PostType;
-import com.spring.blackcat.code.Role;
-import com.spring.blackcat.code.TattooType;
+import com.spring.blackcat.common.code.PostType;
+import com.spring.blackcat.common.code.Role;
+import com.spring.blackcat.common.code.TattooType;
 import com.spring.blackcat.likes.dto.LikesPostResDto;
 import com.spring.blackcat.likes.dto.LikesUserResDto;
 import com.spring.blackcat.post.Post;
