+++ src/test/java/com/spring/blackcat/likes/LikesTest.java
@@ -4,8 +4,8 @@ import com.spring.blackcat.address.Address;
 import com.spring.blackcat.address.AddressRepository;
 import com.spring.blackcat.category.Category;
 import com.spring.blackcat.category.CategoryRepository;
-import com.spring.blackcat.code.Role;
-import com.spring.blackcat.code.TattooType;
+import com.spring.blackcat.common.code.Role;
+import com.spring.blackcat.common.code.TattooType;
 import com.spring.blackcat.post.Post;
 import com.spring.blackcat.post.PostRepository;
 import com.spring.blackcat.tattoo.Tattoo;
