+++ src/test/java/com/spring/blackcat/magazine/MagazineTest.java
